fix: fix linting and spacing issues
