+++ account-kit/rn-signer/example/android/app/src/main/java/accountkit/reactnativesigner/example/MainActivity.kt
@@ -4,9 +4,14 @@ import com.facebook.react.ReactActivity
 import com.facebook.react.ReactActivityDelegate
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.fabricEnabled
 import com.facebook.react.defaults.DefaultReactActivityDelegate
+import android.os.Bundle
 
 class MainActivity : ReactActivity() {
 
+  // Necessary for react-native-screens
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(null)
+  }
   /**
    * Returns the name of the main component registered from JavaScript. This is used to schedule
    * rendering of the component.
