+++ account-kit/rn-signer/example/package.json
@@ -11,10 +11,14 @@
   },
   "dependencies": {
     "@account-kit/signer": "^4.3.0",
+    "@react-navigation/native": "^7.0.3",
+    "@react-navigation/native-stack": "^7.1.0",
     "react": "18.3.1",
     "react-native": "0.76.1",
     "react-native-config": "^1.5.3",
     "react-native-mmkv": "^3.1.0",
+    "react-native-safe-area-context": "^4.14.0",
+    "react-native-screens": "^4.2.0",
     "util": "^0.12.5",
     "viem": "^2.21.41",
     "zustand": "^5.0.1"
